+++ pycite/pycite.py
@@ -55,34 +55,51 @@ def pubmed_journal(bs4_object):
 
 def pubmed_year_volume_pages(bs4_object, show_doi=False):
     """
-    :param bs4_object: An object created with bs4's BeautifulSoup
-    :param show_doi Boolean to control if DOIs should be included in citations. Defaults to False
+    :param bs4_object:
+    :param show_doi: Boolean to control if DOIs should be included in citations. Defaults to False
     :return: A tuple containing the year, volume, and page numbers
     """
+
     # Split only along ; since we can already obtain the DOI elsewhere.
     dates_vol_pages = re.split("[;]", bs4_object.find_all("span", {'class': 'cit'})[0].text)
     # Split year in the form Y M D along a space, year comes first
     year = re.split(" ", dates_vol_pages[0])[0]
+    # This finds volumes and page numbers if they exist
+    # For example in https://pubmed.ncbi.nlm.nih.gov/18952168/
+    # Volumes and page numbers exist as 18(6):756-64.
+    vols_pg_nos = list(filter(None,[re.findall("\\d+\\(.*\\):\\d+-\\d+", x) for x in dates_vol_pages]))
+    volume = None
+    page_numbers = None
+    if vols_pg_nos:
+        # if vols_pgs_nos, then the first is the vol second page numbers
+        vols_pg_nos = re.split(":", vols_pg_nos[0][0])
+        volume = vols_pg_nos[0]
+        page_numbers = vols_pg_nos[1]
     # doi +/ s.*
     paper_identity = bs4_object.find_all("span", {"class": "citation-doi"})[0].text
     # no_doi = re.sub(".*/(?=[Ss])", "", paper_identity.split()[1])
 
-    return year, remove_newlines(paper_identity) if show_doi else year
+    final_volume_year_page = year, remove_newlines(paper_identity) if show_doi else year
+    if volume:
+        final_volume_year_page = final_volume_year_page +  (volume, page_numbers)
+
+    return final_volume_year_page
 
 
 def pubmed_final_citation(bs4_object, show_doi=False):
     """
     :param bs4_object: An object of class BeautifulSoup
-    :param show_doi Boolean to control if DOIs should be included in citations. Defaults to False
+    :param show_doi: Boolean to control if DOIs should be included in citations. Defaults to False
     :return: A pubmed specific citation
     """
-    # TODO: Add page numbers where applicable
     yrs_vols_pages = pubmed_year_volume_pages(bs4_object, show_doi=show_doi)
     # For some reason, year is returned as a tuple with dupes so
     # Need to get the year "twice" here again
-    year = yrs_vols_pages[0] if show_doi else yrs_vols_pages[0][0]
+    year = yrs_vols_pages[0]
     final_citation = (pubmed_authors(bs4_object) + " " + pubmed_title(bs4_object) + " (" +
                       year + ") " + pubmed_journal(bs4_object))
+    if len(yrs_vols_pages) == 4:
+        final_citation = final_citation + ", " + yrs_vols_pages[2] + ", " + yrs_vols_pages[3]
     if show_doi:
         final_citation = final_citation + " " + yrs_vols_pages[1]
 
@@ -175,7 +192,10 @@ class PyCite(object):
                 # Authors (Year) Title, journal, Volume, pages
                 # TODO: Add page numbers
                 # TODO: Make italics
+                # TODO: Write specific methods for NCBI, break up into simpler functions
                 combined_citation = authors_final + " " + title + " (" + year + ") " + journal + ", " + volume
                 out_file.write(f"{combined_citation}\n")
                 final_citations.append(combined_citation)
         return final_citations
+
+