+++ docs/source/changelog.rst
@@ -6,6 +6,9 @@ pycite's changelog
 
 
 * 
+  PUbMed citations now include page numbers, where applicable. See https://github.com/Nelson-Gon/pycite/issues/2
+
+* 
   For PubMed citations, an optional ``show_doi`` argument was added to control whether dois should exist in the citation.
 
 *